Create discrapy.py
This tool scrapes specific discord channels.  To find the channel ID, right-click on the channel and select copy channel ID.  To obtain your user authorization, it's buried somewhere in the settings.  To note, user ID is specific to your user, you will want to keep this protected in a file such as creds.py.